refactor: organise setgetnodes JS logic into folders
