+++ nodes/web/helpers/setgetnodes/main.js
@@ -1,0 +1,4 @@
+import { setColorAndBgColor } from "./style.js";
+import { LGraphNode, showAlert } from "./utils.js";
+
+export { LGraphNode, setColorAndBgColor, showAlert };
