+++ nodes/web/setgetnodes.js
@@ -1,39 +1,6 @@
 import { app } from "../../../scripts/app.js";
+import { LGraphNode, setColorAndBgColor, showAlert } from "./helpers/setgetnodes/main.js";
 
-function setColorAndBgColor(type) {
-  const colorMap = {
-    MODEL: LGraphCanvas.node_colors.blue,
-    LATENT: LGraphCanvas.node_colors.purple,
-    VAE: LGraphCanvas.node_colors.red,
-    CONDITIONING: LGraphCanvas.node_colors.brown,
-    IMAGE: LGraphCanvas.node_colors.pale_blue,
-    CLIP: LGraphCanvas.node_colors.yellow,
-    FLOAT: LGraphCanvas.node_colors.green,
-    MASK: { color: "#1c5715", bgcolor: "#1f401b" },
-    INT: { color: "#1b4669", bgcolor: "#29699c" },
-    CONTROL_NET: { color: "#156653", bgcolor: "#1c453b" },
-    NOISE: { color: "#2e2e2e", bgcolor: "#242121" },
-    GUIDER: { color: "#3c7878", bgcolor: "#1c453b" },
-    SAMPLER: { color: "#614a4a", bgcolor: "#3b2c2c" },
-    SIGMAS: { color: "#485248", bgcolor: "#272e27" },
-  };
-
-  const colors = colorMap[type];
-  if (colors) {
-    this.color = colors.color;
-    this.bgcolor = colors.bgcolor;
-  }
-}
-const LGraphNode = LiteGraph.LGraphNode;
-
-function showAlert(message) {
-  app.extensionManager.toast.add({
-    severity: "warn",
-    summary: "SIG Get/Set",
-    detail: `${message}. Most likely you're missing custom nodes`,
-    life: 5000,
-  });
-}
 app.registerExtension({
   name: "Signature.SetNode",
   registerCustomNodes() {
